Move SSR code under client
Reasons: server side rendering involves both frontend and backend code.
I had put it under server directory because it is executed by server.
However, SSR actually overlaps more with frontend, and actually is
analog to main.tsx on the frontend.

SSR is not appropriate under Web directory, but could work under client,
with shared React modules with Web, to ensure identical library for
SSR and client side rendering.

The minor issue is that server libraries like Koa would need to be
duplicated for SSR and actual Server code. This is not such a big issue
as duplicated (and potentially conflicting) React and Redux.

Signed-off-by: Andy Shu <andyshuhsin@gmail.com>